refactor(AddressInput): split types and styles, and moved component to its own folder
