+++ src/components/primitives/Forms/AddressInput/AddressInput.styled.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+import { Box } from 'components/primitives/Layout/Box';
+
+export const ClickableIcon = styled(Box)`
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+`;
