+++ src/components/primitives/Forms/AddressInput/index.ts
@@ -1,0 +1,1 @@
+export { AddressInput } from './AddressInput';
