+++ src/components/primitives/Forms/AddressInput/types.ts
@@ -1,0 +1,5 @@
+import { InputProps } from 'components/primitives/Forms/Input';
+
+export interface AddressInputProps extends InputProps<string> {
+  enableENSName?: boolean;
+}
